order api completed

+++ src/main/java/com/example/retail_rocket/repository/OrdersRepository.java
@@ -3,5 +3,5 @@ package com.example.retail_rocket.repository;
 import com.example.retail_rocket.model.Orders;
 import org.springframework.data.repository.CrudRepository;
 
-public interface OrdersRepository extends CrudRepository<Orders,Integer> {
+public interface OrdersRepository extends CrudRepository<Orders,Long> {
 }
